Move ShortcutTranslator to utils/qthelpers.py
